Implemented a virtual table function which helps in dealing with table cells that have a rowSpan or colSpan.
Fixed .range() to work better with merged cells using getVirtTable()

Modified .cell() to use getVirtTable(). When using .cell(x,y), if you select a cell that is part of a merged cell, the entire merged cell will be added to the selector. Previously, you had to know the indices in the table to get what you were looking for.

Added .unMerge() to replace a cell that has a rowSpan or cellSpan with individual cells. I think the name is pretty self-explanitory.